+++ packages/plugin-vue/src/main.ts
@@ -35,7 +35,7 @@ export async function transformMain(
   options: ResolvedOptions,
   pluginContext: TransformPluginContext,
   ssr: boolean,
-  asCustomElement: boolean,
+  customElement: boolean,
   ceChildRecord: Map<string, string>,
 ) {
   const { devServer, isProduction, devToolsEnabled } = options
@@ -77,7 +77,7 @@ export async function transformMain(
     options,
     pluginContext,
     ssr,
-    asCustomElement,
+    customElement,
     ceChildRecord,
     filename,
   )
@@ -94,6 +94,7 @@ export async function transformMain(
       options,
       pluginContext,
       ssr,
+      customElement,
     ))
   }
 
@@ -116,7 +117,7 @@ export async function transformMain(
   const stylesCode = await genStyleCode(
     descriptor,
     pluginContext,
-    asCustomElement,
+    customElement,
     attachedProps,
   )
 
@@ -281,6 +282,7 @@ async function genTemplateCode(
   options: ResolvedOptions,
   pluginContext: PluginContext,
   ssr: boolean,
+  customElement: boolean,
 ) {
   const template = descriptor.template!
   const hasScoped = descriptor.styles.some((style) => style.scoped)
@@ -295,6 +297,7 @@ async function genTemplateCode(
       options,
       pluginContext,
       ssr,
+      customElement,
     )
   } else {
     if (template.src) {
@@ -328,7 +331,7 @@ async function genScriptCode(
   options: ResolvedOptions,
   pluginContext: PluginContext,
   ssr: boolean,
-  asCustomElement: boolean,
+  customElement: boolean,
   ceChildRecord: Map<string, string>,
   filename: string,
 ): Promise<{
@@ -338,7 +341,7 @@ async function genScriptCode(
   let scriptCode = `const ${scriptIdentifier} = {}`
   let map: RawSourceMap | undefined
 
-  const script = resolveScript(descriptor, options, ssr)
+  const script = resolveScript(descriptor, options, ssr, customElement)
   if (script) {
     // If the script is js/ts and has no external src, it can be directly placed
     // in the main module.
@@ -377,7 +380,7 @@ async function genScriptCode(
     }
   }
 
-  if (asCustomElement) {
+  if (customElement) {
     scriptCode = transformCEImportSFC(
       scriptCode,
       options,
@@ -395,7 +398,7 @@ async function genScriptCode(
 async function genStyleCode(
   descriptor: SFCDescriptor,
   pluginContext: PluginContext,
-  asCustomElement: boolean,
+  customElement: boolean,
   attachedProps: [string, string][],
 ) {
   let stylesCode = ``
@@ -420,12 +423,12 @@ async function genStyleCode(
           ? `&src=${descriptor.id}`
           : '&src=true'
         : ''
-      const directQuery = asCustomElement ? `&inline` : ``
+      const directQuery = customElement ? `&inline` : ``
       const scopedQuery = style.scoped ? `&scoped=${descriptor.id}` : ``
       const query = `?vue&type=style&index=${i}${srcQuery}${directQuery}${scopedQuery}`
       const styleRequest = src + query + attrsQuery
       if (style.module) {
-        if (asCustomElement) {
+        if (customElement) {
           throw new Error(
             `<style module> is not supported in custom elements mode.`,
           )
@@ -438,7 +441,7 @@ async function genStyleCode(
         stylesCode += importCode
         Object.assign((cssModulesMap ||= {}), nameMap)
       } else {
-        if (asCustomElement) {
+        if (customElement) {
           stylesCode += `\nimport _style_${i} from ${JSON.stringify(
             styleRequest,
           )}`
@@ -448,7 +451,7 @@ async function genStyleCode(
       }
       // TODO SSR critical CSS collection
     }
-    if (asCustomElement) {
+    if (customElement) {
       attachedProps.push([
         `styles`,
         `[${descriptor.styles.map((_, i) => `_style_${i}`).join(',')}]`,